requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
drf-spectacular[sidecar]>=0.27.0
python-decouple>=3.8
Pillow>=10.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
whitenoise>=6.0.0
dj-database-url>=2.0.0

# Development dependencies
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0